kourses/requirements.txt
pillow==12.0.0
python-dotenv==1.2.1
psycopg2==2.9.11
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
stripe==13.2.0
celery==5.5.3
django-celery-beat==2.8.1
redis==7.0.0
django-filter==25.2
drf-spectacular==0.29.0
django-cors-headers==4.9.0
flake8==7.3.0